ASoC: intel/sdw-utils: move soundwire machine driver soc ops
Move Intel SoundWire generic machine driver soc ops to common place
so that it can be used by other platform machine driver.

Link: https://github.com/thesofproject/linux/pull/5068
Link: https://patch.msgid.link/20240801091446.10457-8-Vijendar.Mukunda@amd.com
